+++ FreeMat2/libs/libCore/XML.cpp
@@ -92,7 +92,7 @@ static Array QDomNodeToArray(QDomNode p) {
 
 static Array QDomNamedNodeMapToArray(QDomNamedNodeMap attr) {
   ArrayVector nodes;
-  for (int i=0;i<attr.length();i++)
+  for (size_t i=0;i<attr.length();i++)
     nodes << QDomNodeToArray(attr.item(i));
   return Array::cellConstructor(ArrayMatrix() << nodes);
 }
@@ -198,7 +198,7 @@ void URLRetriever::run()
   if (!m_url.userName().isEmpty())
     m_http.setUser(m_url.userName(), m_url.password());
   m_httpGetId = m_http.get(m_url.path(), m_file);
-  QTimer::singleShot(m_timeout, &m_http, SLOT(abort()));
+  QTimer::singleShot((int)m_timeout, &m_http, SLOT(abort()));
   connect(&m_http, SIGNAL(done(bool)), &m_event, SLOT(quit()));
   connect(&m_http, SIGNAL(requestFinished(int, bool)), this, SLOT(requestFinished(int, bool)));
   m_event.exec();
